MA-23585-2 Fix build issue on lunch aosp_trout_arm64-nxp_stable-userdebug
Add to imx_camera_default/imx_image_process_defaults to defaults to fix below issue:
03:24:57 error: vendor/nxp-opensource/imx/camera/Android.bp:110:1: module "JpegEncTest" variant "android_vendor_arm64_armv8-a_cortex-a53": depends on disabled module "libimageprocess"
03:24:57 error: vendor/nxp-opensource/imx/camera/Android.bp:110:1: module "JpegEncTest" variant "android_vendor_arm64_armv8-a_cortex-a53": depends on disabled module "libimxcamerahwl_impl"
03:24:57 error: vendor/nxp-opensource/imx/camera/Android.bp:110:1: module "JpegEncTest" variant "android_vendor_arm64_armv8-a_cortex-a53": depends on disabled module "libimageprocess"
03:24:57 error: vendor/nxp-opensource/imx/camera/Android.bp:110:1: module "JpegEncTest" variant "android_vendor_arm64_armv8-a_cortex-a53": depends on disabled module "libimxcamerahwl_impl"

Also fix a build warning due to compare int and uint32_t.

Reviewed-on: http://androidsource.nxp.com/project/28794
(cherry picked from commit 7443c0ca88a24a96b2796b23c839821c7c12fe5a)
Reviewed-on: http://androidsource.nxp.com/project/28929

+++ camera/JpegEncTest.cpp
@@ -80,7 +80,7 @@ int main(int argc, char *argv[]) {
         }
 
         ret = fread(inBuf, 1, inSize, pSrcFileStream);
-        if (ret != inSize) {
+        if (ret != (int)inSize) {
             printf("Failed to read file %s for %d bytes\n", fileSrcName, inSize);
             ret = -1;
             goto exit;
